fix freespace util for linux
fr this time

+++ dlls/util.cpp
@@ -3129,6 +3129,7 @@ uint64_t getFreeSpace(const std::string& path) {
 		ALERT(at_console, "Error getting free space.\n");
 		return 0;
 	}
-	return stat.f_bavail * stat.f_bsize;
+
+	return (uint64_t)stat.f_bavail * (uint64_t)stat.f_bsize;
 #endif
 }
